refactor(util): extract fs helpers and stream JSON export
Move `tempFilePath` and `renameWithRetry` from `src/util.ts` into a
focused `src/util/fs.ts`, alongside a new `writeAtomically(path, write)`
helper that captures the temp-file + rename + best-effort cleanup
pattern previously copy-pasted into the telemetry writer and
`CliCredentialManager.atomicWriteFile`. Both call sites now use the
shared helper; the SSH config / support bundle / CLI manager keep their
bespoke flows and just update imports.

Rewrite `writeJsonArrayExport` to stream chunks through
`Readable.from(...)` into `createWriteStream` via `stream/promises`
`pipeline`. Memory stays flat regardless of `maxTotalBytes`, which has
no enforced ceiling and can comfortably exceed the default 100 MB cap.

Tests for `tempFilePath` and `renameWithRetry` move into
`test/unit/util/fs.test.ts` alongside a new `writeAtomically` suite
covering success, partial-write rollback, and callback return values.
