+++ src/util.ts
@@ -164,51 +164,6 @@ export function countSubstring(needle: string, haystack: string): number {
 	return count;
 }
 
-const transientRenameCodes: ReadonlySet<string> = new Set([
-	"EPERM",
-	"EACCES",
-	"EBUSY",
-]);
-
-/**
- * Rename with retry for transient Windows filesystem errors (EPERM, EACCES,
- * EBUSY). On Windows, antivirus, Search Indexer, cloud sync, or concurrent
- * processes can briefly lock files causing renames to fail.
- *
- * On non-Windows platforms, calls renameFn directly with no retry.
- *
- * Matches the strategy used by VS Code (pfs.ts) and graceful-fs: 60s
- * wall-clock timeout with linear backoff (10ms increments) capped at 100ms.
- */
-export async function renameWithRetry(
-	renameFn: (src: string, dest: string) => Promise<void>,
-	source: string,
-	destination: string,
-	timeoutMs = 60_000,
-	delayCapMs = 100,
-): Promise<void> {
-	if (process.platform !== "win32") {
-		return renameFn(source, destination);
-	}
-	const startTime = Date.now();
-	for (let attempt = 1; ; attempt++) {
-		try {
-			return await renameFn(source, destination);
-		} catch (err) {
-			const code = (err as NodeJS.ErrnoException).code;
-			if (
-				!code ||
-				!transientRenameCodes.has(code) ||
-				Date.now() - startTime >= timeoutMs
-			) {
-				throw err;
-			}
-			const delay = Math.min(delayCapMs, attempt * 10);
-			await new Promise((resolve) => setTimeout(resolve, delay));
-		}
-	}
-}
-
 /**
  * Wraps `arg` in `"..."` unless every character is in the shell-safe
  * whitelist (matching Python `shlex.quote`'s set: alphanumerics plus
@@ -239,13 +194,4 @@ export function escapeShellArg(arg: string): string {
 		return `"${escaped}"`;
 	}
 	return `'${arg.replace(/'/g, "'\\''")}'`;
-}
-
-/**
- * Generate a temporary file path by appending a suffix with a random component.
- * The suffix describes the purpose of the temp file (e.g. "temp", "old").
- * Example: tempFilePath("/a/b", "temp") → "/a/b.temp-k7x3f9qw"
- */
-export function tempFilePath(basePath: string, suffix: string): string {
-	return `${basePath}.${suffix}-${crypto.randomUUID().substring(0, 8)}`;
 }
