+++ test/unit/core/supportBundleLogs.test.ts
@@ -5,14 +5,14 @@ import * as path from "node:path";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 import { appendVsCodeLogs } from "@/core/supportBundleLogs";
-import { renameWithRetry } from "@/util";
+import { renameWithRetry } from "@/util/fs";
 
 import { createMockLogger } from "../../mocks/testHelpers";
 
 // Wrap renameWithRetry so individual tests can override it via
 // mockRejectedValueOnce; by default it calls through to the real impl.
-vi.mock("@/util", async () => {
-	const actual = await vi.importActual<typeof import("@/util")>("@/util");
+vi.mock("@/util/fs", async () => {
+	const actual = await vi.importActual<typeof import("@/util/fs")>("@/util/fs");
 	return { ...actual, renameWithRetry: vi.fn(actual.renameWithRetry) };
 });
